Pass old inputs to Read during Refresh
These changes expose a resource's old inputs to the provider when
performing a refresh. The provider can take advantage of this to
maintain the accuracy of any additional data or metadata in the
resource's inputs that may need to be updated during the refresh.

This is required for the complete implementation of
https://github.com/pulumi/pulumi-terraform/pull/349. Without access to
the old inputs for a resource, TF-based providers would lose all
information about default population during a refresh.