Handle older parent resources
When an older pulumi/pulumi is used for a parent than that used for a chid resoruce, it was possible for the `__aliases` property to be undefined.

These changes handle that case, and make the recently added private properties of Resource optional to better represent the fact that instances of Resource may not have these values set.