Set recurse=True on serializer
Dill by default does not recurisvely capture global state, which leads to dangling name references.  For many dynamic provider use cases it is necessary to be able to correctly capture globals.

Also, customize the pickler to ensure dicts are sorted prior to serializing so that we get a more deterministic serialization.